refactor: 페이징 관련 구조 통합
* feat: 컨트롤러 limit, page 요청 통일

* Feat: 쿼리 대문자로 변경

* refactor: 쿠폰 관련 필요없는 주석 제거

* feat: 레슨 관련 페이징 통합 및 구조 정리

* refactor: 쿼리 대소문자 수정

* feat: 프로필 관련 페이징 구조 통일

* feat: 리뷰 관련 n+1 문제 해결 -> profile 쪽 같이 join 통해

* feat: 레슨 관련 N + 1 문제 해결

+++ src/main/java/com/threestar/trainus/domain/lesson/teacher/mapper/LessonApplicationMapper.java
@@ -30,7 +30,7 @@ public class LessonApplicationMapper {
 
 	//신청자 목록과 전체 개수를 리스트 응답 DTO로 변환
 	public static LessonApplicationListResponseDto toListResponseDto(
-		List<LessonApplication> applications, Long totalCount) {
+		List<LessonApplication> applications, int totalCount) {
 
 		// 각 신청을 응답 DTO로 변환
 		List<LessonApplicationResponseDto> applicationDtos = applications.stream()
@@ -40,7 +40,7 @@ public class LessonApplicationMapper {
 		// 리스트 응답 DTO 생성
 		return LessonApplicationListResponseDto.builder()
 			.lessonApplications(applicationDtos)
-			.count(totalCount.intValue())
+			.count(totalCount)
 			.build();
 	}
